+++ mesh_protocol_dwm1001-dev/src/MessageHandler.c
@@ -131,13 +131,13 @@ void MessageHandler_SendPing(Node node) {
   // transmit message via driver
   Driver_TransmitPing(node, msg);
 
-  int8_t currentSlot = TimeKeeping_CalculateCurrentSlotNum(node);
+  uint8_t currentSlot = TimeKeeping_CalculateCurrentSlotNum(node);
   bool isOwn = SlotMap_IsOwnSlot(node, currentSlot);
   bool isPending = SlotMap_IsPendingSlot(node, currentSlot);
 
   // if the current slot is not already pending or own, add it to pending (it was a new reservation attempt)
   if (!isOwn && !isPending) {
-    int8_t neighbors[MAX_NUM_NODES - 1];
+    uint8_t neighbors[MAX_NUM_NODES - 1];
 
     // get the neighbors of the node at this particular time, cause these neighbors need to acknowledge the ping 
     // Note: currently only the number of neighbors that ack'ed is checked, not which neighbors exactly (simpler implementation)
@@ -188,20 +188,20 @@ static void joinNetwork(Node node, Message msg) {
   TimeKeeping_SetFrameStartTimeForLastPreamble(node, msg);
 
   // release possible own and pending slots when joining a new network, because these are not valid anymore
-  int8_t ownSlotsBuffer[NUM_SLOTS];
-  int8_t numOwn = SlotMap_GetOwnSlots(node, &ownSlotsBuffer[0], NUM_SLOTS);
-  for (int i = 0; i < numOwn; ++i) {
+  uint8_t ownSlotsBuffer[NUM_SLOTS];
+  uint8_t numOwn = SlotMap_GetOwnSlots(node, &ownSlotsBuffer[0], NUM_SLOTS);
+  for (uint8_t i = 0; i < numOwn; ++i) {
     SlotMap_ReleaseOwnSlot(node, ownSlotsBuffer[i]);
   };
 
-  int8_t pendingSlotsBuffer[NUM_SLOTS];
-  int8_t numPending = SlotMap_GetPendingSlots(node, &pendingSlotsBuffer[0], NUM_SLOTS);
-  for (int i = 0; i < numPending; ++i) {
+  uint8_t pendingSlotsBuffer[NUM_SLOTS];
+  uint8_t numPending = SlotMap_GetPendingSlots(node, &pendingSlotsBuffer[0], NUM_SLOTS);
+  for (uint8_t i = 0; i < numPending; ++i) {
     SlotMap_ReleasePendingSlot(node, pendingSlotsBuffer[i]);
   };
 
 #if DEBUG_VERBOSE
-  printf("%d: Node %d joined network %d \n", (int) localTime, node->id, msg->networkId);
+  printf("%d: Node %d joined network %d \n", (uint8_t) localTime, node->id, msg->networkId);
 #endif
 };
 
@@ -210,26 +210,26 @@ static void updateSlots(Node node, Message msg) {
   SlotMap_UpdatePendingSlotAcks(node, msg);
   
   // add pending slots that were acknowledged by all required nodes to own slots
-  int8_t acknowledgedPendingSlots[MAX_NUM_PENDING_SLOTS];
-  int8_t numAcked = SlotMap_GetAcknowledgedPendingSlots(node, &acknowledgedPendingSlots[0], MAX_NUM_PENDING_SLOTS);
-  for(int i = 0; i < numAcked; ++i) {
+  uint8_t acknowledgedPendingSlots[MAX_NUM_PENDING_SLOTS];
+  uint8_t numAcked = SlotMap_GetAcknowledgedPendingSlots(node, &acknowledgedPendingSlots[0], MAX_NUM_PENDING_SLOTS);
+  for(uint8_t i = 0; i < numAcked; ++i) {
     SlotMap_ChangePendingToOwn(node, acknowledgedPendingSlots[i]);
   };
   
   // check if own slots were reported as colliding by the sending node
-  int8_t collidingOwnSlots[MAX_NUM_OWN_SLOTS];
-  int8_t numCollidingOwn = SlotMap_CheckOwnSlotsForCollisions(node, msg, &collidingOwnSlots[0], MAX_NUM_OWN_SLOTS);
+  uint8_t collidingOwnSlots[MAX_NUM_OWN_SLOTS];
+  uint8_t numCollidingOwn = SlotMap_CheckOwnSlotsForCollisions(node, msg, &collidingOwnSlots[0], MAX_NUM_OWN_SLOTS);
 
   // release colliding own slots
-  for (int i = 0; i < numCollidingOwn; ++i) {
+  for (uint8_t i = 0; i < numCollidingOwn; ++i) {
     SlotMap_ReleaseOwnSlot(node, collidingOwnSlots[i]);
   };
 
   // check if pending slots were reported as colliding by the sending node
-  int8_t collidingPendingSlots[MAX_NUM_PENDING_SLOTS];
-  int8_t numCollidingPending = SlotMap_CheckPendingSlotsForCollisions(node, msg, &collidingPendingSlots[0], MAX_NUM_PENDING_SLOTS);
+  uint8_t collidingPendingSlots[MAX_NUM_PENDING_SLOTS];
+  uint8_t numCollidingPending = SlotMap_CheckPendingSlotsForCollisions(node, msg, &collidingPendingSlots[0], MAX_NUM_PENDING_SLOTS);
   // release colliding pending slots
-  for (int i = 0; i < numCollidingPending; ++i) {
+  for (uint8_t i = 0; i < numCollidingPending; ++i) {
     SlotMap_ReleasePendingSlot(node, collidingPendingSlots[i]);
   };
 
@@ -241,7 +241,7 @@ static void updateSlots(Node node, Message msg) {
   SlotMap_UpdateThreeHopSlotMap(node, msg);
 
   // cancel the next schedule if the corresponding slot was released 
-  int8_t nextScheduledSlot = Scheduler_GetSlotOfNextSchedule(node);
+  uint8_t nextScheduledSlot = Scheduler_GetSlotOfNextSchedule(node);
   int16_t idxCollidingOwn = Util_Int8tArrayFindElement(&collidingOwnSlots[0], nextScheduledSlot, numCollidingOwn);
   int16_t idxCollidingPending = Util_Int8tArrayFindElement(&collidingPendingSlots[0], nextScheduledSlot, numCollidingPending);
